requirements.txt
torch
torchvision
pandas
matplotlib
numpy
streamlit
Pillow
